feat: move data export logic to it's own file
